+++ openapi-ts.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from '@hey-api/openapi-ts';
 export default defineConfig({
   client: '@hey-api/client-fetch',
   input: './build/swagger.json',
-  output: `../narrowcasting-client/src/api/`,
+  output: `../aurora-client/src/api/`,
   plugins: [
     '@hey-api/schemas',
     '@hey-api/services',
