+++ src/modules/handlers/screen/index.ts
@@ -1,7 +1,7 @@
 export { default as CurrentlyPlayingTrackHandler } from './curently-playing-track-handler';
 export { default as CenturionScreenHandler } from './centurion-screen-handler';
-export { default as GewisPosterScreenHandler } from './poster/gewis-poster-screen-handler';
-export { default as HubblePosterScreenHandler } from './poster/hubble-poster-screen-handler';
+export { default as GewisPosterScreenHandler } from './poster/gewis/gewis-poster-screen-handler';
+export { default as HubblePosterScreenHandler } from './poster/hubble/hubble-poster-screen-handler';
 export { default as RoomResponsibleLegacyHandler } from './room-responsible-legacy-handler';
 export { default as StageEffectsHandler } from './stage-effects-handler';
 export { default as TimeTrailRaceScreenHandler } from './time-trail-race-screen-handler';
