+++ src/modules/handlers/screen/poster/gewis/gewis-poster-service.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosError } from 'axios';
-import { HttpApiException } from '../../../../helpers/custom-error';
+import { HttpApiException } from '../../../../../helpers/custom-error';
 
 export interface GEWISPhotoAlbumParams {
   albumIds: number[];
