+++ src/modules/handlers/screen/poster/hubble/hubble-poster-screen-handler.ts
@@ -1,4 +1,4 @@
-import BasePosterScreenHandler from "./base-poster-screen-handler";
+import BasePosterScreenHandler from "../base-poster-screen-handler";
 import { Namespace } from "socket.io";
 
 export default class HubblePosterScreenHandler extends BasePosterScreenHandler {
